requirements.txt
numpy>=1.21.0
pandas>=1.3.0
optuna>=3.0.0
requests==2.31.0
tqdm>=4.60.0
cython>=0.29.0
psutil>=5.8.0
ta>=0.10.0
Flask==2.3.3
Flask-CORS==4.0.0
Flask-RESTful==0.3.10
authlib==1.2.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
six==1.16.0
werkzeug==2.3.7
gunicorn==21.2.0
matplotlib==3.10.5
scipy==1.16.1
Flask-Compress==1.18
APScheduler==3.11.0
Flask-SQLAlchemy==2.5.1
